+++ config/book_setup.rb
@@ -1,7 +1,7 @@
 module GPTK
   class Book
-    num_chapters = 13 # BE EXTREMELY CAREFUL SETTING THIS VALUE!!!
-    chapter_fragments = 5 # This determines how many chapter fragments are in a chapter. Larger values = more words.
+    num_chapters = 12 # BE EXTREMELY CAREFUL SETTING THIS VALUE!!!
+    chapter_fragments = 7 # This determines how many chapter fragments are in a chapter. Larger values = more words.
     chapter_fragment_words = 3000 # How large each chapter fragment is
     CONFIG = {
       num_chapters: num_chapters,
@@ -12,6 +12,7 @@ module GPTK
       prompt: "For the chapter title and content, refer EXPLICITLY to the outline, and if included, the prior chapter summary and current chapter summary. Refer to your context for memory of prior content, as well. Chapter title should be an H1 element SPECIFICALLY (# character in markdown) followed by the chapter name. Chapter titles must match those in the outline EXACTLY. Generate AT LEAST #{chapter_fragment_words} words.".freeze,
       post_prompt: 'Make SURE to include the chapter number with the chapter title.',
       command_code: 'The response should FIRST contain the chapter content, THEN, delineated with 3 dashes (markdown horizontal line), a summary of the current chapter fragment. Delineation of the summary MUST be 3 dashes SPECIFICALLY.'.freeze,
+      meta_prompt: 'Maintain continuity and do NOT repeat any previously generated material. Generate as much content as possible. Put any conversational material or description from yourself in brackets so as to separate it from the content you generate. AVOID expose, i.e. "telling" instead of "showing" in between fragments. AVOID explaining what is going on at the end of a fragment/chapter, such as "the stage was set" or "the chapter closed". AVOID repeating phrases or elements that have already been used, such as "shivers went down her spine" and "the air was thick with tension". AVOID cliches and trite writing style.'.freeze,
       # Define atomic parsing steps via Ruby procs, so they can be passed individually to parser methods
       # Each element is composed of 1) A regular expression search pattern, and 2) A string replacement pattern
       default_parsers: { # Regex, replacement